Добавил Unit-тест

+++ proj_1OPO/Cylinder.h
@@ -6,7 +6,7 @@ private:
     int baseX, baseY, height, baseRadius;
 
 public:
-    Cylinder(double d, const std::string& o, int bx, int by, int h, int br);
+    Cylinder(double d, const std::string& o, double bx, double by, double h, double br);
     void print() const override;
     bool matchesCondition(const std::string& condition) const override;
 };
